converter working and tested
