+++ orcid-message-listener/src/main/java/org/orcid/listener/ReIndexListener.java
@@ -21,7 +21,10 @@ import java.util.Map;
 import javax.annotation.Resource;
 
 import org.orcid.jaxb.model.message.OrcidProfile;
+import org.orcid.jaxb.model.record_rc3.Record;
+import org.orcid.listener.clients.LockedRecordException;
 import org.orcid.listener.clients.Orcid12APIClient;
+import org.orcid.listener.clients.Orcid20APIClient;
 import org.orcid.listener.clients.S3Updater;
 import org.orcid.listener.clients.SolrIndexUpdater;
 import org.orcid.utils.listener.LastModifiedMessage;
@@ -39,6 +42,8 @@ public class ReIndexListener {
     @Resource
     private Orcid12APIClient orcid12ApiClient;
     @Resource
+    private Orcid20APIClient orcid20ApiClient;
+    @Resource
     private SolrIndexUpdater solrIndexUpdater;
     @Resource
     private S3Updater s3Updater; 
@@ -52,8 +57,14 @@ public class ReIndexListener {
     public void processMessage(final Map<String, String> map) {
         LastModifiedMessage message = new LastModifiedMessage(map);
         LOG.info("Recieved " + MessageConstants.Queues.REINDEX + " message for orcid " + message.getOrcid() + " " + message.getLastUpdated());
-        OrcidProfile profile = orcid12ApiClient.fetchPublicProfile(message.getOrcid());
-        solrIndexUpdater.updateSolrIndex(profile);
-        s3Updater.updateS3(profile);
+        try{
+            OrcidProfile profile = orcid12ApiClient.fetchPublicProfile(message.getOrcid());
+            Record record = orcid20ApiClient.fetchPublicProfile(message.getOrcid());//can we not just transform the above?
+            solrIndexUpdater.updateSolrIndex(record);
+            s3Updater.updateS3(profile);            
+        }catch (LockedRecordException e){
+            //if the record is locked then 'blank' it in Solr.
+            solrIndexUpdater.updateSolrIndexForLockedRecord(message.getOrcid(),message.getLastUpdated());
+        }
     }
 }
