+++ orcid-message-listener/src/main/java/org/orcid/listener/clients/LockedRecordException.java
@@ -1,0 +1,5 @@
+package org.orcid.listener.clients;
+
+public class LockedRecordException extends Exception{
+
+}
