+++ orcid-message-listener/src/main/java/org/orcid/listener/clients/Orcid20APIClient.java
@@ -19,36 +19,33 @@ package org.orcid.listener.clients;
 import java.net.URI;
 import java.net.URISyntaxException;
 
+import javax.annotation.Resource;
 import javax.ws.rs.core.MediaType;
 
-import org.orcid.jaxb.model.message.OrcidMessage;
-import org.orcid.jaxb.model.message.OrcidProfile;
+import org.orcid.jaxb.model.record_rc3.Record;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import com.fasterxml.jackson.jaxrs.json.JacksonJaxbJsonProvider;
 import com.sun.jersey.api.client.Client;
 import com.sun.jersey.api.client.ClientResponse;
 import com.sun.jersey.api.client.WebResource;
-import com.sun.jersey.api.client.config.ClientConfig;
-import com.sun.jersey.api.client.config.DefaultClientConfig;
 
 @Component
-public class Orcid12APIClient {
+public class Orcid20APIClient {
 
-    Logger LOG = LoggerFactory.getLogger(Orcid12APIClient.class);
-    protected final Client jerseyClient;
+    Logger LOG = LoggerFactory.getLogger(Orcid20APIClient.class);
+    
+    @Resource
+    protected Client jerseyClient;
+    
     protected final URI baseUri;
 
     @Autowired
-    public Orcid12APIClient(@Value("${org.orcid.message-lisener.api12BaseURI}") String baseUri) throws URISyntaxException {
-        LOG.info("Creating Orcid12APIClient with baseUri = " + baseUri);
-        ClientConfig config = new DefaultClientConfig();
-        config.getClasses().add(JacksonJaxbJsonProvider.class);
-        jerseyClient = Client.create(config);
+    public Orcid20APIClient(@Value("${org.orcid.message-lisener.api20BaseURI}") String baseUri) throws URISyntaxException {
+        LOG.info("Creating Orcid20APIClient with baseUri = " + baseUri);
         this.baseUri = new URI(baseUri);
     }
 
@@ -57,17 +54,15 @@ public class Orcid12APIClient {
      * 
      * @param orcid
      * @return
-     * @throws LockedRecordException 
      */
-    public OrcidProfile fetchPublicProfile(String orcid) throws LockedRecordException {
+    public Record fetchPublicProfile(String orcid) throws LockedRecordException{
         WebResource webResource = jerseyClient.resource(baseUri);
-        ClientResponse response = webResource.path(orcid + "/orcid-profile").accept(MediaType.APPLICATION_XML).get(ClientResponse.class);
+        ClientResponse response = webResource.path(orcid + "/record").accept(MediaType.APPLICATION_XML).get(ClientResponse.class);
         if (response.getStatus() != 200) {
             if (response.getStatus() == 409)
                 throw new LockedRecordException();
             throw new RuntimeException("Failed : HTTP error code : " + response.getStatus());
         }
-        OrcidMessage output = response.getEntity(OrcidMessage.class);
-        return output.getOrcidProfile();
+        return response.getEntity(Record.class);
     }
 }
