+++ orcid-message-listener/src/main/java/org/orcid/listener/listeners/updated/UpdatedOrcidWorker.java
@@ -21,7 +21,10 @@ import java.util.Date;
 import javax.annotation.Resource;
 
 import org.orcid.jaxb.model.message.OrcidProfile;
+import org.orcid.jaxb.model.record_rc3.Record;
+import org.orcid.listener.clients.LockedRecordException;
 import org.orcid.listener.clients.Orcid12APIClient;
+import org.orcid.listener.clients.Orcid20APIClient;
 import org.orcid.listener.clients.S3Updater;
 import org.orcid.listener.clients.SolrIndexUpdater;
 import org.orcid.utils.listener.LastModifiedMessage;
@@ -40,6 +43,8 @@ public class UpdatedOrcidWorker implements RemovalListener<String, LastModifiedM
     @Resource
     private Orcid12APIClient orcid12ApiClient;
     @Resource
+    private Orcid20APIClient orcid20ApiClient;
+    @Resource
     private SolrIndexUpdater solrIndexUpdater;
     @Resource
     private S3Updater s3Updater;
@@ -54,14 +59,24 @@ public class UpdatedOrcidWorker implements RemovalListener<String, LastModifiedM
             LastModifiedMessage m = removal.getValue();
             LOG.info("Removing " + removal.getKey() + " from UpdatedOrcidCacheQueue");
             LOG.info("Processing " + m.getLastUpdated());
-            OrcidProfile profile = orcid12ApiClient.fetchPublicProfile(m.getOrcid());
-            Date lastModifiedFromprofile = profile.getOrcidHistory().getLastModifiedDate().getValue().toGregorianCalendar().getTime();
-            Date lastModifiedFromSolr = solrIndexUpdater.retrieveLastModified(m.getOrcid());
-            // note this is slightly different from existing behaviour
-            if (lastModifiedFromprofile.after(lastModifiedFromSolr))
-                solrIndexUpdater.updateSolrIndex(profile);
-            // now do the same for S3...
-            s3Updater.updateS3(profile);
+            try{
+                OrcidProfile profile = orcid12ApiClient.fetchPublicProfile(m.getOrcid());
+                Record record = orcid20ApiClient.fetchPublicProfile(m.getOrcid());//can we not just transform the above?
+                Date lastModifiedFromprofile = profile.getOrcidHistory().getLastModifiedDate().getValue().toGregorianCalendar().getTime();
+                Date lastModifiedFromSolr = solrIndexUpdater.retrieveLastModified(m.getOrcid());
+                // note this is slightly different from existing behaviour
+                if (lastModifiedFromprofile.after(lastModifiedFromSolr))
+                    solrIndexUpdater.updateSolrIndex(record);
+                // now do the same for S3...
+                s3Updater.updateS3(profile);
+            }catch(LockedRecordException e){
+                //if the record is locked then 'blank' it in Solr.
+                Date lastModifiedFromSolr = solrIndexUpdater.retrieveLastModified(m.getOrcid());                // note this is slightly different from existing behaviour
+                if (m.getLastUpdated().after(lastModifiedFromSolr))
+                    solrIndexUpdater.updateSolrIndexForLockedRecord(m.getOrcid(), m.getLastUpdated());
+                //TODO: something with S3
+            }
+
         }
     }
 }
