+++ orcid-message-listener/src/main/resources/message-listener.properties
@@ -18,9 +18,10 @@
 org.orcid.listener.lastUpdateSecondsToWait=30
 org.orcid.listener.lastUpdateForceCleanup=true
 org.orcid.message-lisener.api12BaseURI=http://localhost:8080/orcid-pub-web/v1.2/
++org.orcid.message-lisener.api20BaseURI=https://localhost:8443/orcid-pub-web/v2.0_rc3/
 org.orcid.message-lisener.writeToFileNotS3=true
 
-org.orcid.core.indexPublicProfile=true
+org.orcid.core.indexPublicProfile=false
 
 org.orcid.persistence.solr.url=http://localhost:8080/orcid-solr-web
 org.orcid.persistence.solr.read.only.url=http://localhost:8080/orcid-solr-web
