+++ orcid-model/src/main/java/org/orcid/jaxb/model/record_rc1/WorkExternalIdentifierType.java
@@ -96,7 +96,8 @@ public enum WorkExternalIdentifierType implements Serializable {
                 return wit;
             }
         }
-        throw new IllegalArgumentException(value);
+        return WorkExternalIdentifierType.OTHER_ID;
+        //throw new IllegalArgumentException(value);
     }
 
 }
